refactor(blogs): Refactor and optimize blogs query and components
- Updated `getBlogs.ts` to include a `status` field in the `BlogPostAPI` interface.
- Modified query parameters in `useGetBlogs` to default `status` to 'draft'.
- Updated mapping logic to include `status` in the data transformation.
- rename BlogListView folder to more appropriate name
