+++ app/content-panel/blogs/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import BlogListView from '../../../components/module-components/create-blogs/BlogListView';
+import BlogListView from '../../../components/module-components/blogs/BlogListView';
 
 const page = () => {
   return <BlogListView />;
