+++ components/module-components/blogs/BlogListView.module.scss
@@ -2,7 +2,7 @@
 
 .mainContainer {
   width: 100%;
-  max-width: 1200px;
+  max-width: 1400px;
   padding: 20px;
   margin: 0 auto;
